Do not promote disks twice
This patch fixes an issue where disks were being promoted
twice due to a stale controller cache on systems with heavy
load. The disk promotion reconciler depended upon the promo
task reference to be written to the VM status, but upon
re-entering reconcile, the status.taskRef field may be empty
if the cache is not updated.

This fix removes status.taskRef and instead fetches all of a
VM's recent tasks at the beginning of each reconcile.